Renounce Religion Completed; Changed Item Grid Row Attack Attribute symbol to the universal one; but show details on the row details; Support Affiliation level changes; Support for re-affiliating.

+++ Rogue.NET.Core/Model/Scenario/Alteration/AttackAttribute.cs
@@ -8,7 +8,7 @@ namespace Rogue.NET.Core.Model.Scenario.Alteration
     {
         public double Attack { get; set; }
         public double Resistance { get; set; }
-        public double Weakness { get; set; }
+        public int Weakness { get; set; }
 
         public AttackAttribute()
         {
